refactor: move ingredients to shared kernel
