+++ production/src/main/scala/dev/atedeg/mdm/production/Types.scala
@@ -74,12 +74,3 @@ final case class QuintalsOfIngredient(quintals: WeightInQuintals, ingredient: In
  */
 final case class WeightInQuintals(n: PositiveDecimal) derives Times
 
-/**
- * An ingredient that may be needed by a [[Recipe recipe]] to produce a [[CheeseType type of cheese]].
- */
-enum Ingredient:
-  case Milk extends Ingredient
-  case Cream extends Ingredient
-  case Rennet extends Ingredient
-  case Salt extends Ingredient
-  case Probiotics extends Ingredient
