+++ products-shared-kernel/src/main/scala/dev/atedeg/mdm/products/Ingredients.scala
@@ -1,0 +1,11 @@
+package dev.atedeg.mdm.products
+
+/**
+ * An ingredient that may be needed by a [[Recipe recipe]] to produce a [[CheeseType type of cheese]].
+ */
+enum Ingredient:
+  case Milk extends Ingredient
+  case Cream extends Ingredient
+  case Rennet extends Ingredient
+  case Salt extends Ingredient
+  case Probiotics extends Ingredient
